% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_att.R
\name{get_att}
\alias{get_att}
\alias{get_att.character}
\alias{get_att.NetCDF}
\alias{get_att.ZarrGroup}
\title{Get Zarr Attribute}
\usage{
get_att(z, var, att)

\method{get_att}{character}(z, var, att)

\method{get_att}{NetCDF}(z, var, att)

\method{get_att}{ZarrGroup}(z, var, att)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_nz}}

\item{var}{integer or character zero-based index id of variable of interest
or name of variable of interest. -1 or "global" for global attributes}

\item{att}{integer or character zero-based index id of attribute of interest
or name of attribute of interest.}
}
\value{
vector containing requested attribute \link[RNetCDF]{att.get.nc}
}
\description{
Get Zarr Attribute
}
\examples{

z <- open_nz(z_demo())

get_att(z, -1, 2)

get_att(z, "global", "Conventions")

get_att(z, 0, 3)

get_att(z, "latitude", "long_name")

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {
  nc <- z_demo(format = "netcdf")

  (get_att(RNetCDF::open.nc(nc), 0, 0))

  (get_att(RNetCDF::open.nc(nc), "global", 1))
}

}
