% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_var}
\alias{get_var}
\alias{get_var.character}
\alias{get_var.NetCDF}
\alias{get_var.ZarrGroup}
\alias{get_var.NULL}
\title{Get Zarr Variable}
\usage{
get_var(z, var, start = NA, count = NA, collapse = TRUE, unpack = FALSE, ...)

\method{get_var}{character}(z, var, start = NA, count = NA, collapse = TRUE, unpack = FALSE, ...)

\method{get_var}{NetCDF}(z, var, start = NA, count = NA, collapse = TRUE, unpack = FALSE, ...)

\method{get_var}{ZarrGroup}(z, var, start = NA, count = NA, collapse = TRUE, unpack = FALSE, ...)

\method{get_var}{`NULL`}(z, var, start = NA, count = NA, collapse = TRUE, unpack = FALSE, ...)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_nz}}

\item{var}{integer or character zero-based index id of variable of interest
or name of variable of interest.}

\item{start}{integer vector with length equal to the number of dimensions of var.
Uses R-style 1 indexing. If NA the entire array is returned.}

\item{count}{integer vector with length equal to the number of dimensions of var.
Specifies the size of the returned array along the dimension in question. Can not
be NA if start is not NA. NA count in one or more dimensions can be used to
indicate all of a given dimension.}

\item{collapse}{logical if TRUE degenerated dimensions (length=1) will be omitted.}

\item{unpack}{logical if TRUE and the variable carries `scale_factor` and/or
`add_offset` attributes, the returned values are unpacked as
`(value * scale_factor) + add_offset`.}

\item{...}{passed to RNetCDF var.get.nc}
}
\value{
array of data
}
\description{
Get Zarr Variable
}
\examples{

if(requireNamespace("pizzarr", quietly = TRUE)) {
z <- open_nz(z_demo())

latitude <- get_var(z, 0)

class(latitude)
dim(latitude)

pr <- get_var(z, "pr") |>
  aperm(c(3,2,1))

pr[pr > 1000] <- NA

image(pr[,,1], col = hcl.colors(12, "PuBuGn", rev = TRUE),
      useRaster = TRUE, axes = FALSE)

dim(pr)

# subsetting

pr <- get_var(z, "pr", start = c(1, 1, 1), count = c(5, 5, -1))

dim(pr)

}

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {
  nc <- z_demo(format = "netcdf")

  (get_var(nc, 0))

  pr <- get_var(nc, "pr")

  image(pr[,,1], col = hcl.colors(12, "PuBuGn", rev = TRUE),
        useRaster = TRUE, axes = FALSE)

}

}
