% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_att.R
\name{inq_att}
\alias{inq_att}
\alias{inq_att.character}
\alias{inq_att.NetCDF}
\alias{inq_att.ZarrGroup}
\title{Inquire Zarr Attribute}
\usage{
inq_att(z, var, att)

\method{inq_att}{character}(z, var, att)

\method{inq_att}{NetCDF}(z, var, att)

\method{inq_att}{ZarrGroup}(z, var, att)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_nz}}

\item{var}{integer or character zero-based index id of variable of interest
or name of variable of interest. -1 or "global" for global attributes}

\item{att}{integer or character zero-based index id of attribute of interest
or name of attribute of interest.}
}
\value{
list similar to that returned by \link[RNetCDF]{att.inq.nc}
}
\description{
Inquire Zarr Attribute
}
\examples{

z <- open_nz(z_demo())

inq_att(z, -1, 0)

inq_att(z, "global", "Conventions")

inq_att(z, 0, 3)

inq_att(z, "latitude", "long_name")

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {
  nc <- z_demo(format = "netcdf")

  (inq_att(RNetCDF::open.nc(nc), 0, 0))
}
}
