% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_grp.R
\name{inq_grp}
\alias{inq_grp}
\title{Inquire Zarr Group}
\usage{
inq_grp(z, group = "/")
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_zarr}}

\item{group}{NOTUSED}
}
\value{
list similar to that returned by grp.inq.nc
}
\description{
NOTE: only functions on a root group thus far
}
\examples{

z <- open_zarr(system.file("extdata", "bcsd_obs_1999.zarr", package = "rnz"))

inq_grp(z)

# equivalent data in NetCDF
requireNamespace("RNetCDF", quietly = TRUE) {
  nc <- system.file("extdata", "bcsd_obs_1999.nc", package = "rnz")

  (RNetCDF::grp.inq.nc(RNetCDF::open.nc(nc)))
}

}
