% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_store.R
\name{inq_nz_source}
\alias{inq_nz_source}
\title{Inquire Zarr Store}
\usage{
inq_nz_source(z)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_nz}}
}
\value{
list similar to that returned by \link[RNetCDF]{file.inq.nc}
}
\description{
Inquire Zarr Store
}
\examples{

z <- open_nz(z_demo())

inq_nz_source(z)

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {
  nc <- system.file("extdata", "bcsd_obs_1999.nc", package = "rnz")

  (RNetCDF::file.inq.nc(RNetCDF::open.nc(nc)))
}
}
