% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_store.R
\name{inq_store}
\alias{inq_store}
\title{Inquire Zarr Store}
\usage{
inq_store(z)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_zarr}}
}
\value{
list similar to that returned by file.inq.nc
}
\description{
Inquire Zarr Store
}
\examples{

z <- open_zarr(system.file("extdata", "bcsd_obs_1999.zarr", package = "rnz"))

zarr_inq_store(z)

# equivalent data in NetCDF
requireNamespace("RNetCDF", quietly = TRUE) {
  nc <- system.file("extdata", "bcsd_obs_1999.nc", package = "rnz")

  (RNetCDF::file.inq.nc(RNetCDF::open.nc(nc)))
}

}
