% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{nzdump}
\alias{nzdump}
\alias{nzdump.character}
\alias{nzdump.NetCDF}
\alias{nzdump.ZarrGroup}
\title{Dump zarr like ncdump}
\usage{
nzdump(nz)

\method{nzdump}{character}(nz)

\method{nzdump}{NetCDF}(nz)

\method{nzdump}{ZarrGroup}(nz)
}
\arguments{
\item{nz}{path to store, open zarr store, or open zarr group}
}
\value{
a list of lines printed to the console invisibly
}
\description{
Prints an ncdump-like output to the console.
}
\examples{
store <- rnz::z_demo()

nzdump(store)

nc <- z_demo(format = "netcdf")

nzdump(nz)

}
