% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z_demo}
\alias{z_demo}
\title{nz demo data directory}
\usage{
z_demo(format = "zarr")
}
\arguments{
\item{format}{character "netcdf" or "zarr"}
}
\value{
netcdf sample data or unzips and returns a demo Zarr store directory
}
\description{
Data derived from: https://gdo-dcp.ucllnl.org/downscaled_cmip_projections/
}
\examples{
list.files(z_demo(format = "zarr"), recursive = TRUE, all.files = TRUE)

basename(z_demo(format = "netcdf"))
}
