% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z_dir}
\alias{z_dir}
\title{Zarr demo data directory
unzips and returns a demo Zarr store directory.}
\usage{
z_dir()
}
\description{
Zarr demo data directory
unzips and returns a demo Zarr store directory.
}
\examples{
list.files(z_dir(), recursive = TRUE, all.files = TRUE)

}
