% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{zdump}
\alias{zdump}
\title{Dump zarr like ncdump}
\usage{
zdump(store)
}
\arguments{
\item{store}{path to store, open zarr store, or open zarr group}
}
\value{
a list of lines printed to the console invisibly
}
\description{
Prints an ncdump-like output to the console.
}
\examples{
store <- rnz::z_demo()

zdump(store)

}
