% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close.R
\name{close_nz}
\alias{close_nz}
\alias{close_nz.NetCDF}
\alias{close_nz.ZarrGroup}
\alias{close_nz.NULL}
\title{close netcdf or zarr
Closes handle. Doesn't currently do anything for zarr stores.}
\usage{
close_nz(z)

\method{close_nz}{NetCDF}(z)

\method{close_nz}{ZarrGroup}(z)

\method{close_nz}{`NULL`}(z)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_nz}}
}
\description{
close netcdf or zarr
Closes handle. Doesn't currently do anything for zarr stores.
}
\examples{
if(requireNamespace("pizzarr", quietly = TRUE)) {

z <- z_demo()

bcsd <- open_nz(z)

close_nz(bcsd)

}

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {
  nc <- z_demo(format = "netcdf")

  bcsd <- open_nz(nc)

  close_nz(bcsd)
}
}
