% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_var}
\alias{get_var}
\title{Get Zarr Variable}
\usage{
get_var(z, var)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_zarr}}

\item{var}{integer or character zero-based index id of variable of interest
or name of variable of interest.}
}
\value{
array of data
}
\description{
Get Zarr Variable
}
\examples{

z <- open_zarr(system.file("extdata", "bcsd_obs_1999.zarr", package = "rnz"))

latitude <- get_var(z, 0)

class(latitude)

dim(latitude)

pr <- get_var(z, "pr")

dim(pr)

# equivalent data in NetCDF
requireNamespace("RNetCDF", quietly = TRUE) {
  nc <- system.file("extdata", "bcsd_obs_1999.nc", package = "rnz")

  (RNetCDF::var.get.nc(RNetCDF::open.nc(nc), 0))
}

}
