% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_att.R
\name{inq_att}
\alias{inq_att}
\title{Inquire Zarr Attribute}
\usage{
inq_att(z, var, att)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_zarr}}

\item{var}{integer zero-based index id of variable of interest. If missing
or -1 for global.}

\item{att}{integer zero-based index id of attribute of interest}
}
\value{
list similar to that returned by \link[RNetCDF]{att.inq.nc}
}
\description{
Inquire Zarr Attribute
}
\examples{

z <- open_zarr(system.file("extdata", "bcsd_obs_1999.zarr", package = "rnz"))

inq_att(z, att = 0)

inq_att(z, -1, 0)

inq_att(z, 0, 0)

# equivalent data in NetCDF
requireNamespace("RNetCDF", quietly = TRUE) {
  nc <- system.file("extdata", "bcsd_obs_1999.nc", package = "rnz")

  (RNetCDF::att.inq.nc(RNetCDF::open.nc(nc), 0, 0))
}

}
