% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_dim.R
\name{inq_dim}
\alias{inq_dim}
\title{Inquire Zarr Dimension}
\usage{
inq_dim(z, dim)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_zarr}}

\item{dim}{integer zero-based index id of dimension of interest}
}
\value{
list similar to that returned by \link[RNetCDF]{dim.inq.nc}
}
\description{
NOTE: assumes the `_ARRAY_DIMENSION` convention from `xarray`
}
\examples{

z <- open_zarr(z_dir())

inq_dim(z, 0)

inq_dim(z, "latitude")

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {
  nc <- system.file("extdata", "bcsd_obs_1999.nc", package = "rnz")

  (RNetCDF::dim.inq.nc(RNetCDF::open.nc(nc), 0))
}

}
