% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_dim.R
\name{inq_dim}
\alias{inq_dim}
\alias{inq_dim.character}
\alias{inq_dim.NetCDF}
\alias{inq_dim.ZarrGroup}
\alias{inq_dim.NULL}
\title{Inquire Zarr Dimension}
\usage{
inq_dim(z, dim)

\method{inq_dim}{character}(z, dim)

\method{inq_dim}{NetCDF}(z, dim)

\method{inq_dim}{ZarrGroup}(z, dim)

\method{inq_dim}{`NULL`}(z, dim)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_nz}}

\item{dim}{integer zero-based index id of dimension of interest}
}
\value{
list similar to that returned by \link[RNetCDF]{dim.inq.nc}
}
\description{
NOTE: assumes the `_ARRAY_DIMENSION` convention from `xarray`
}
\examples{

z <- open_nz(z_demo())

inq_dim(z, 0)

inq_dim(z, "latitude")

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {
  nc <- z_demo(format = "netcdf")

  (inq_dim(nc, 0))
}
}
