% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_grp.R
\name{inq_grp}
\alias{inq_grp}
\alias{inq_grp.character}
\alias{inq_grp.NetCDF}
\alias{inq_grp.ZarrGroup}
\alias{inq_grp.NULL}
\title{Inquire Zarr Group}
\usage{
inq_grp(z, group = "/")

\method{inq_grp}{character}(z, group = "/")

\method{inq_grp}{NetCDF}(z, group = "/")

\method{inq_grp}{ZarrGroup}(z, group = "/")

\method{inq_grp}{`NULL`}(z, group = "/")
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_nz}}

\item{group}{NOTUSED}
}
\value{
list similar to that returned by \link[RNetCDF]{grp.inq.nc}
}
\description{
NOTE: only functions on a root group thus far
}
\examples{

z <- open_nz(z_demo())

inq_grp(z)

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {#'
  inq_grp(z_demo(format = "netcdf"))
}
}
