% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_var.R
\name{inq_var}
\alias{inq_var}
\title{Inquire Zarr Variable}
\usage{
inq_var(z, var)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_zarr}}

\item{var}{integer or character zero-based index id of variable of interest
or name of variable of interest.}
}
\value{
list similar to that returned by \link[RNetCDF]{var.inq.nc}
}
\description{
Inquire Zarr Variable
}
\examples{

z <- open_zarr(system.file("extdata", "bcsd_obs_1999.zarr", package = "rnz"))

inq_var(z, 0)

# equivalent data in NetCDF
requireNamespace("RNetCDF", quietly = TRUE) {
  nc <- system.file("extdata", "bcsd_obs_1999.nc", package = "rnz")

  (RNetCDF::var.inq.nc(RNetCDF::open.nc(nc), 0))
}

}
