% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inq_var.R
\name{inq_var}
\alias{inq_var}
\alias{inq_var.character}
\alias{inq_var.NetCDF}
\alias{inq_var.ZarrGroup}
\alias{inq_var.NULL}
\title{Inquire Zarr Variable}
\usage{
inq_var(z, var)

\method{inq_var}{character}(z, var)

\method{inq_var}{NetCDF}(z, var)

\method{inq_var}{ZarrGroup}(z, var)

\method{inq_var}{`NULL`}(z, var)
}
\arguments{
\item{z}{an open ZarrGroup as returned by \link{open_nz}}

\item{var}{integer or character zero-based index id of variable of interest
or name of variable of interest.}
}
\value{
list similar to that returned by \link[RNetCDF]{var.inq.nc}
}
\description{
Inquire Zarr Variable
}
\examples{

z <- open_nz(z_demo())

inq_var(z, 0)

inq_var(z, "pr")

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {
  nc <- z_demo(format = "netcdf")
  nc <- rnz::open_nz(nc, backend = "RNetCDF")
  inq_var(nc, 0)
}
}
