% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_zarr}
\alias{open_zarr}
\title{open zarr}
\usage{
open_zarr(store)
}
\arguments{
\item{store}{a pizzarr store or path to a zarr store}
}
\value{
ZarrGroup
}
\description{
open zarr
}
\examples{
z <- z_dir()

bcsd <- open_zarr(z)

class(bcsd)

zarr <- pizzarr::DirectoryStore$new(z)

class(zarr)

bcsd <- open_zarr(zarr)

class(bcsd)

# equivalent data in NetCDF
if(requireNamespace("RNetCDF", quietly = TRUE)) {
  nc <- system.file("extdata", "bcsd_obs_1999.nc", package = "rnz")

  (RNetCDF::open.nc(nc))
}

}
