% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z_demo}
\alias{z_demo}
\title{Zarr demo data directory
unzips and returns a demo Zarr store directory.}
\usage{
z_demo()
}
\description{
Zarr demo data directory
unzips and returns a demo Zarr store directory.
}
\examples{
list.files(z_demo(), recursive = TRUE, all.files = TRUE)

}
